feat: Add get_date.bat function for retrieving formatted dates
This commit adds a new batch file, get_date.bat, which is used to retrieve formatted dates in batch scripts. The function accepts two optional parameters: the date offset (default is -1 for yesterday) and the date format (default is yyyyMMdd). It uses PowerShell to calculate the date and store it in the RESULT variable.

The get_date.bat file is added to the Functions directory.